requirements.txt
requests>=2.28.0
pytest>=7.0.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
groq>=1.0.0